reorder files in directories
