+++ detox/src/devices/drivers/AndroidAppiumDriver.js
@@ -1,0 +1,86 @@
+/**
+ * Created by Or Evron on 07/08/2018
+ */
+const AppiumDriverBase = require("./AppiumDriverBase");
+const DetoxApi = require("../../android/espressoapi/Detox");
+const EspressoDetoxApi = require("../../android/espressoapi/EspressoDetox");
+const AAPT = require("../android/AAPT");
+const {interruptProcess} = require("../../utils/exec");
+const path = require('path');
+
+class AndroidAppiumDriver extends AppiumDriverBase {
+    constructor(client) {
+        super(client);
+        client = client.client;
+        this.expect = require("../../android/expect");
+        this.expect.setInvocationManager(this.invocationManager);
+        this.aapt = new AAPT();
+        this._desiredCapabilities = Object.assign(
+            client.configuration.appium.desiredCapabilities || {},
+            {
+                allowTestPackages: true,
+                clearSystemFiles: true,
+                app: client.configuration.appium.desiredCapabilities.wrapperApp,
+                otherApps: JSON.stringify([client.configuration.appium.desiredCapabilities.app, client.configuration.appium.desiredCapabilities.testBinaryPath]),
+                optionalIntentArguments: `--es detoxServer ${client.configuration.server} --es detoxSessionId ${client.configuration.sessionId} --es packageName ${client.configuration.appium.desiredCapabilities.bundleId}.test`,
+                appActivity: "MainActivity"
+            });
+    }
+
+    async getBundleIdFromBinary(binaryPath) {
+        return this._desiredCapabilities.bundleId;
+    }
+
+    async uninstallApp(deviceId, bundleId) {
+        await this._driver.removeAppFromDevice(`${this._desiredCapabilities.bundleId}`);
+    }
+
+    async installApp(deviceId, binaryPath, testBinaryPath) {
+        await this._driver.installApp(JSON.parse(this._desiredCapabilities.otherApps)[0]);
+        await this._driver.launchApp();
+    }
+
+    async deliverPayload(params) {
+        if (params.url) {
+            const call = DetoxApi.startActivityFromUrl(params.url);
+            await this.invocationManager.execute(call);
+        }
+    }
+
+    async setURLBlacklist(urlList) {
+        const call = EspressoDetoxApi.setURLBlacklist(urlList);
+        await this.invocationManager.execute(call);
+    }
+
+    async enableSynchronization() {
+        const call = EspressoDetoxApi.setSynchronization(true);
+        await this.invocationManager.execute(call);
+    }
+
+    async disableSynchronization() {
+        const call = EspressoDetoxApi.setSynchronization(false);
+        await this.invocationManager.execute(call);
+    }
+
+    async cleanup(deviceId, bundleId) {
+        await this._driver.quit();
+        await this.terminateInstrumentation();
+    }
+
+    async terminateInstrumentation() {
+        if (this.instrumentationProcess) {
+            await interruptProcess(this.instrumentationProcess);
+            this.instrumentationProcess = null;
+        }
+    }
+
+    defaultLaunchArgsPrefix() {
+        return "-e ";
+    }
+
+    exportGlobals() {
+        this.expect.exportGlobals(this._driver);
+    }
+}
+
+module.exports = AndroidAppiumDriver;
